refactor: reorganize structure
