+++ packages/playground/src/components/debug-menu.ts
@@ -15,12 +15,11 @@ import '@shoelace-style/shoelace/dist/components/color-picker/color-picker.js';
 import {
   createEvent,
   type FrameBlockModel,
-  getModelsByRange,
+  getCurrentBlockRange,
   MouseMode,
   NonShadowLitElement,
-  SelectionUtils,
   ShapeMouseMode,
-  updateSelectedTextType,
+  updateBlockType,
 } from '@blocksuite/blocks';
 import type { EditorContainer } from '@blocksuite/editor';
 import {
@@ -150,16 +149,30 @@ export class DebugMenu extends NonShadowLitElement {
   ) {
     e.preventDefault();
     this.blockTypeDropdown.hide();
-
-    updateSelectedTextType('affine:list', listType);
+    const blockRange = getCurrentBlockRange(this.page);
+    if (!blockRange) {
+      return;
+    }
+    const models =
+      blockRange.startModel === blockRange.endModel
+        ? [blockRange.startModel]
+        : [
+            blockRange.startModel,
+            ...blockRange.betweenModels,
+            blockRange.endModel,
+          ];
+    updateBlockType(models, 'affine:list', listType);
   }
 
   private _addCodeBlock(e: PointerEvent) {
     e.preventDefault();
     this.blockTypeDropdown.hide();
 
-    const range = SelectionUtils.getCurrentRange();
-    const startModel = getModelsByRange(range)[0];
+    const blockRange = getCurrentBlockRange(this.page);
+    if (!blockRange) {
+      throw new Error("Can't add code block without a selection");
+    }
+    const startModel = blockRange.startModel;
     const parent = this.page.getParent(startModel);
     const index = parent?.children.indexOf(startModel);
     const blockProps = {
@@ -176,7 +189,19 @@ export class DebugMenu extends NonShadowLitElement {
     e.preventDefault();
     this.blockTypeDropdown.hide();
 
-    updateSelectedTextType('affine:paragraph', type);
+    const blockRange = getCurrentBlockRange(this.page);
+    if (!blockRange) {
+      return;
+    }
+    const models =
+      blockRange.startModel === blockRange.endModel
+        ? [blockRange.startModel]
+        : [
+            blockRange.startModel,
+            ...blockRange.betweenModels,
+            blockRange.endModel,
+          ];
+    updateBlockType(models, 'affine:paragraph', type);
   }
 
   private _switchEditorMode() {
